cambiado el bigdecimal por double en price

+++ src/main/java/product/BedlamProducts.java
@@ -1,7 +1,5 @@
 package product;
 
-import java.math.BigDecimal;
-
 public interface BedlamProducts {
-    BigDecimal getPrice();
+    Double getPrice();
 }
