+++ src/main/java/product/Product.java
@@ -3,17 +3,15 @@ package product;
 import java.math.BigDecimal;
 
 public class Product {
-    private BigDecimal price;
+    private Double price;
     private String name;
 
 
-
-
-    public Product(BigDecimal price, String name){
+    public Product(Double price, String name){
     this.price = price;
     this.name = name;}
 
-        public BigDecimal getPrice() {return price;}
+        public Double getPrice() {return price;}
 
         public void setPrice() {this.price = price;}
